Performance optimization
Moved all into one component again
Applied useMemo and useCallback functions to minimize rerendering
