+++ src/components/EmailDetail/Messages/MessagesOverview.tsx
@@ -3,7 +3,7 @@ import CircularProgress from '@mui/material/CircularProgress'
 import EmailDetailOptions from './EmailDetailOptions'
 import DraftMessage from './DisplayVariants/DraftMessage'
 import ReadUnreadMessage from './DisplayVariants/ReadUnreadMessage'
-import ComposeEmail from '../../Compose/ComposeEmailContainer'
+import ComposeEmail from '../../Compose/ComposeEmail'
 import * as local from '../../../constants/emailDetailConstants'
 import * as global from '../../../constants/globalConstants'
 import * as draft from '../../../constants/draftConstants'
